moved the class out and added support for language stuff (#1)
